Move eta, html-minifier, lodash and other dependencies outside the server bundle
